requirements.txt
langchain
openai
huggingface_hub
python-dotenv
faiss-cpu
youtube-transcript-api
streamlit